requirements.txt
openai
langchain
faiss-cpu
sentence-transformers
pdfplumber
python-docx
streamlit
fastapi
uvicorn
pydantic
python-dotenv
